+++ lib/src/post_activation/fields_auto_fill.dart
@@ -1,0 +1,1 @@
+enum FieldsAutoFill { None, Fields, FieldsAndSetters }